endorse-backend/requirements.txt
# c:\Users\HomePC\endorse\endorse-backend\requirements.txt
fastapi
uvicorn[standard]
cloudinary
python-dotenv
requests
